requirements.txt
aiodns
aioredis
argon2-cffi
asyncpg
cchardet
cython
discord.py[voice]
gtts
jishaku
lavalink
matplotlib
numpy
Pillow
pycryptodomex
pygit2
PyLyrics
pypng
pyqrcode
python-dateutil
sympy
wheel
